+++ tests/test_core.py
@@ -7,13 +7,6 @@ from numpy import testing as npt
 import ixmp
 from ixmp.default_path_constants import CONFIG_PATH
 
-from testing_utils import (
-    test_mp,
-    test_mp_props,
-    test_mp_use_default_dbprops_file,
-    test_mp_use_db_config_path,
-)
-
 test_args = ('Douglas Adams', 'Hitchhiker')
 can_args = ('canning problem', 'standard')
 
@@ -253,3 +246,25 @@ def test_meta(test_mp):
     obs = scen.get_meta('test_string')
     exp = test_dict['test_string']
     assert obs == exp
+
+
+def test_load_scenario_data(test_mp):
+    scen = test_mp.Scenario(*can_args, cache=True)
+    scen.load_scenario_data()
+    assert ('par', 'd') in scen._pycache  # key exists
+    df = scen.par('d', filters={'i': ['seattle']})
+    obs = df.loc[0, 'unit']
+    exp = 'km'
+    assert obs == exp
+
+
+def test_load_scenario_data_clear_cache(test_mp):
+    # this fails on commit: 4376f54
+    scen = test_mp.Scenario(*can_args, cache=True)
+    scen.load_scenario_data()
+    scen.clear_cache(name='d')
+
+
+def test_load_scenario_data_raises(test_mp):
+    scen = test_mp.Scenario(*can_args, cache=False)
+    pytest.raises(ValueError, scen.load_scenario_data)