+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 
 # Next Release
+
+- [#52](https://github.com/iiasa/ixmp/pull/52): Add ability to load all scenario data into memory for fast subsequent computation.